requirements.txt
# Pic2Doc
# Python Dependencies
# Installation: pip install -r requirements.txt

# Word document generation
python-docx==1.1.0

# Excel file processing
openpyxl==3.1.2

# Image processing (for orientation detection)
Pillow>=10.0.0

# Python Version Requirement
# Python >= 3.8
